+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## [Unreleased]
-- **[Fixed]** Machine now only waits for the remaining time instead of always waiting the fix amount.
+- **[Added]** `behavior_machine.visualization` which has function `visualize_behavior_machine` that can visualize any machine and it's subcomponents.
+- **[Fixed]** Machine now only waits for the remaining time instead of always waiting the fixed amount.
 
 ## [0.3.3] - 2021-05-03 
 - **[Added]** `SaveFlowState`, `SetFlowFromBoardState`,`SetBoardState`,`GetBoardState` and `SetFlowState`. Three states that help with board access.